requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
sqlalchemy
python-dotenv
jinja2
pdfkit
pytest